awscdk-packages/requirements.txt
aws-cdk-lib==2.175.1
constructs>=10.0.0,<11.0.0
pyyaml==6.0.1
aws-cdk.lambda-layer-kubectl-v31==2.0.0